requirements.txt
attrs==25.3.0
aws-cdk-lib==2.219.0
aws-cdk.asset-awscli-v1==2.2.242
aws-cdk.asset-node-proxy-agent-v6==2.1.0
aws-cdk.aws-lambda-python-alpha==2.219.0a0
aws-cdk.cloud-assembly-schema==48.12.0
aws_lambda_powertools==3.20.0
boto3==1.40.44
botocore==1.40.44
cattrs==25.2.0
colorama==0.4.6
constructs==10.4.2
importlib_resources==6.5.2
iniconfig==2.1.0
jmespath==1.0.1
jsii==1.115.0
packaging==25.0
pluggy==1.6.0
publication==0.0.3
Pygments==2.19.2
pytest==8.4.2
python-dateutil==2.9.0.post0
s3transfer==0.14.0
six==1.17.0
typeguard==2.13.3
typing_extensions==4.15.0
urllib3==2.5.0
